# java/quarkus/swiftref/src/main/openapi/swiftref.yaml
openapi: 3.0.0
info:
  title: SWIFTRef API
  description: |
    One of SWIFT's latest delivery channels for SWIFTRef reference data are Application Programming Interfaces (APIs). APIs offer financial institutions, corporates and third party service providers an automated data look-up service for real-time identification and validation of critical payments reference data within the SWIFTRef utility.
    
    **v4.0.0 Release notes**

    * HTTP request header `Accept` now used for the customer expected format (JSON/XML):
      * it should only be `application/json` or `application/xml`;
      * if an unexpected accept header is received, an error 406 is returned. If not sent, the request is accepted and the default (JSON) will be applied.
    * Errors - all the V4 APIs return the error in the same format:
      * a collection, whose structure is `[ {errorObject1}, {errorObject2}, ... ]`;
      * `more_info` error attribute became optional.
    * All the APIs within the OpenAPI contract have the same version. The version corresponds to the major version of the OpenAPI contract (OpenAPI contract 4.0.0 only contain V4 APIs).
    * Session cookie removed.
    * Header `API-Status` (API verbose) removed.
    * Response customization (EOL, colons) removed:
      * JSON responses are not returned pretty-printed, but raw.
    * OpenAPI changes
      * standardizations as per Swagger HUB rules
      * RegExp validations for each input field added
...

...
    
...

...

...

...

...
servers:
...
security:
  - oauthBearerToken: [ ]
paths:
...
  '/ibans/{iban}':
    get:
      tags:
        - ibans
      summary: Get details for an IBAN.
      description: For a given IBAN, you can obtain the components of the IBAN.
      operationId: getIbanDetails
      parameters:
        - in: path
          name: iban
          schema:
            $ref: '#/components/schemas/Iban'
          required: true
          description: IBAN for which details are requested
      responses:
        '200':
          description: Corresponding IBAN found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IbanDetailsWithIsoFlag'
            application/xml:
              schema:
                $ref: '#/components/schemas/IbanDetailsWithIsoFlag'
        '401':
          $ref: '#/components/responses/Unallowed'
        '404':
          description: No corresponding IBAN found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
              example:
                errors:
                  - code: REDA.API.INVI
                    severity: Fatal
                    text: Invalid IBAN
                    user_message: Invalid IBAN
            application/xml:
              schema:
                $ref: '#/components/schemas/Errors'
              example:
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <errors>
                <error>
                <code>REDA.API.INVI</code>
                <severity>Fatal</severity>
                <text>Invalid IBAN</text>
                <user_message>Invalid IBAN</user_message>
                </error>
                </errors>
        '405':
          $ref: '#/components/responses/InvalidHttpMethod'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  '/bics/{bic}':
    get:
      tags:
        - bics
      summary: Get details of a BIC.
      description: For a given BIC, you can retrieve the details of that BIC, that is the bank name or business name, and the address and the SWIFT service details.
      operationId: getBicDetails
      parameters:
        - in: path
          name: bic
          schema:
            $ref: '#/components/schemas/Bic'
          required: true
          description: BIC (8-characters or 11-characters) for which details are requested
      responses:
        '200':
          description: Corresponding BIC found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BicIsoDetails'
            application/xml:
              schema:
                $ref: '#/components/schemas/BicIsoDetails'
        '401':
          $ref: '#/components/responses/Unallowed'
        '404':
          description: BIC not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errors'
              example:
                errors:
                  - code: REDA.API.BINF
                    severity: Fatal
                    text: The supplied BIC does not exist within the BIC Directory
                    user_message: No corresponding BIC found
            application/xml:
              schema:
                $ref: '#/components/schemas/Errors'
              example:
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <errors>
                <error>
                <code>REDA.API.BINF</code>
                <severity>Fatal</severity>
                <text>The supplied BIC does not exist within the BIC Directory</text>
                <user_message>No corresponding BIC found</user_message>
                </error>
                </errors>
        '405':
          $ref: '#/components/responses/InvalidHttpMethod'
        '406':
          $ref: '#/components/responses/NotAcceptable'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    oauthBearerToken:
      type: http
      scheme: bearer
      bearerFormat: opaque OAuth 2.0
      description: |
        Use the access token obtained as a result of OAuth 2.0 flows. 
        
        SWIFT supports two OAuth grant types depending on the API service.
        * JWT-Bearer grant type [RFC 7523](https://tools.ietf.org/html/rfc7523)
        * Password grant type
        
        This API uses password grant type.
        
        Please visit [SWIFT OAuth Token API](https://developer.swift.com/swift-oauth-token-api) for more information and examples.
        
        In this declaration only the basic security element to transport the bearer token of an OAuth2 process is declared.
...
  responses:
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
        application/xml:
          schema:
            $ref: '#/components/schemas/Errors'
    InvalidHttpMethod:
      description: Request HTTP method is not accepted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            errors:
              - code: REDA.API.IREQ
                severity: Fatal
                text: Request performed with HTTP method other than GET
                user_message: Invalid request
    NotAcceptable:
      description: The server does not support the format asked by the consumer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            errors:
              - code: REDA.API.NOAC
                severity: Fatal
                text: Requested output format is not supported
    Unallowed:
      description: You do not have sufficient privileges to execute this service
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            errors:
              - code: REDA.API.ILIC
                severity: Fatal
                text: You do not have sufficient privileges to execute this service
                user_message: You do not have sufficient privileges to execute this service
  schemas:
...
    StringMax16:
      description: Generic string definition, from 1 to 16 characters long
      type: string
      minLength: 1
      maxLength: 16
...
    StringMax35:
      description: Generic string definition, from 1 to 35 characters long
      type: string
      minLength: 1
      maxLength: 35
    StringMax70:
      description: Generic string definition, from 1 to 70 characters long
      type: string
      minLength: 1
      maxLength: 70
    StringMax105:
      description: Generic string definition, from 1 to 105 characters long
      type: string
      minLength: 1
      maxLength: 105
    StringMax140:
      description: Generic string definition, from 1 to 140 characters long
      type: string
      minLength: 1
      maxLength: 140
    StringMax255:
      description: Generic string definition, from 1 to 255 characters long
      type: string
      minLength: 1
      maxLength: 255
...
    TwoDigitCode:
      description: Generic 2-digit code
      type: string
      pattern: '^[0-9]{2}$'
...
    AlphaNumStringMax30:
      description: Alphanumeric string definition, from 1 to 30 characters long
      type: string
      pattern: '^[0-9a-zA-Z]{1,30}$'
    InstitutionName:
      $ref: '#/components/schemas/StringMax140'
      # description: Name by which an institution is known and which is usually used to identify that institution
    InstitutionDepartment:
      $ref: '#/components/schemas/StringMax70'
...
    CountryName:
      $ref: '#/components/schemas/StringMax70'
...
    IsoPostalCode:
      type: string
      minLength: 1
      maxLength: 16
      description: Identifier consisting of a group of letters and/or numbers that is added to a postal address to assist the sorting of mail
    IsoPostOfficeIdentifier:
      description: Numbered box in a post office, assigned to a person or organisation, where letters are kept until called for
      type: string
      minLength: 1
      maxLength: 16
...
    FinPlusDistinguishedName:
      $ref: '#/components/schemas/StringMax255'
      # description: Unique identifier of an entity that sends or receives messages over SWIFTNet.
      # example: 'ou=shd,o=handsess,o=swift'
...
    Error:
      description: Custom error schema to support detailed error message.
      required:
        - code
        - severity
        - text
      type: object
      properties:
        severity:
          description: Specifies the severity of the error.
          type: string
          enum:
            - Fatal
            - Transient
            - Logic
        code:
          description: Specifies the custom error code as defined by the service provider.
          type: string
          minLength: 3
          maxLength: 70
        text:
          $ref: '#/components/schemas/StringMax255'
          # description: Specifies the detail error message identifying the cause of the error.
        user_message:
          $ref: '#/components/schemas/StringMax255'
          # description: A human-readable text describing the error.
        more_info:
          description: Specifies an URL to find more information about the error.
          type: string
          format: uri
      xml:
        name: error
      additionalProperties: false
      example:
        code: REDA.API.CODE
        severity: Fatal
        text: Error description
        user_message: User-friendly error description
    Errors:
      # Source: "SWIFT API Design Guidelines from Bitbucket" section 9.2
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
          minItems: 1
      xml:
        name: errors
      additionalProperties: false
...
    UnstructuredAddress:
      type: object
      description: ISO-compliant unstructured address
      required:
        - address_line_1
      properties:
        address_line_1:
          $ref: '#/components/schemas/StringMax35'
        address_line_2:
          $ref: '#/components/schemas/StringMax35'
        address_line_3:
          $ref: '#/components/schemas/StringMax35'
      additionalProperties: false
      example:
        address_line_1: "ONE OXFORD CTR"
        address_line_2: "301 GRANT ST, STE 3000"
        address_line_3: "PITTSBURGH PA 15219, US"
    StructuredAddress:
      type: object
      description: ISO-compliant structured address
      required:
        - country_code
        - country_name
      properties:
        street_name:
          $ref: '#/components/schemas/StringMax70'
        building_number:
          $ref: '#/components/schemas/StringMax16'
        building_name:
          $ref: '#/components/schemas/StringMax35'
        post_box:
          $ref: '#/components/schemas/IsoPostOfficeIdentifier'
        floor:
          $ref: '#/components/schemas/StringMax70'
        room:
          $ref: '#/components/schemas/StringMax70'
        post_code:
          $ref: '#/components/schemas/IsoPostalCode'
        town_name:
          $ref: '#/components/schemas/StringMax35'
        town_location_name:
          $ref: '#/components/schemas/StringMax35'
        district_name:
          $ref: '#/components/schemas/StringMax35'
        country_subdivision_name:
          $ref: '#/components/schemas/StringMax35'
        country_subdivision_code:
          $ref: '#/components/schemas/StringMax35'
        country_code:
          $ref: '#/components/schemas/CountryCode'
        country_name:
          $ref: '#/components/schemas/CountryName'
      additionalProperties: false
      example:
        street_name: "GRANT ST"
        building_number: "301"
        building_name: "ONE OXFORD CTR"
        post_box: "PO BOX 867"
        room: "STE 3000"
        post_code: "15219"
        town_name: "PITTSBURGH"
        country_subdivision_name: "PENNSYLVANIA"
        country_subdivision_code: "PA"
        country_code: "US"
        country_name: "UNITED STATES OF AMERICA"
    SwiftSubtypeIndicator:
      type: string
      pattern: '^[0-9A-Z]{4}$'
      description: SWIFT subtype indicator
      example: SUPE
...
    TimeOffset:
      type: string
      pattern: '^[±+-]\d\d:\d\d$'
      description: Difference in hours and minutes between two places
    SwiftService:
      description: SWIFT service code and name
      type: object
      xml:
        name: swift_service
      required:
        - code
        - name
      properties:
        code:
          type: string
          pattern: '^[A-Z]{2}[A-Z+]$'
          description: Code given by the service provider (SWIFT) to uniquely identify a service it offers
        name:
          $ref: '#/components/schemas/StringMax105'
          # description: Name given by the service provider (SWIFT) to uniquely identify a service it offers
      example:
        code: TGT
        name: TARGET2 DIRECT PARTICIPANTS
      additionalProperties: false
    SwiftConnectivity:
      description: Institution connection to SWIFTNet messaging services
      type: object
      required:
        - fin
        - interact
        - fileact
      properties:
        fin:
          type: boolean
          description: True if connected to SWIFT FIN, false otherwise
        finplus:
          type: object
          description: FINplus Distinguished Names (DN), used to uniquely identify an entity that sends or receives messages over SWIFTNet.
          properties:
            future:
              $ref: '#/components/schemas/FinPlusDistinguishedName'
            live:
              $ref: '#/components/schemas/FinPlusDistinguishedName'
            pilot:
              $ref: '#/components/schemas/FinPlusDistinguishedName'
          additionalProperties: false
        interact:
          type: boolean
          description: True if used in at least one service on SWIFTNet InterAct, false otherwise
        fileact:
          type: boolean
          description: True if used in at least one service on SWIFTNet FileAct, false otherwise
      additionalProperties: false
      example:
        fin: true
        finplus:
          future: 'ou=shd,o=handsess,o=swift'
          live: 'ou=shd,o=handsess,o=swift'
          pilot: 'ou=shd,o=handsess,o=swift'
        interact: false
        fileact: false
...
    Iban:
      type: string
      minLength: 5
      maxLength: 34
      pattern: '^[a-zA-Z]{2}[0-9]{2}[a-zA-Z0-9]{1,30}$'
      description: IBAN (International Bank Account Number)
      example: IT60X0542811101000000123456
    Bic:
      type: string
      description: Bank Identifier Code - ISO 9362
      pattern: '^[A-Z]{6}[A-Z2-9][A-NP-Z0-9]([A-Z0-9]{3})?$'
      example: SWHTBEBBXXX
...
    CountryCode:
      type: string
      pattern: '^[a-zA-Z]{2}$'
      description: 2-character ISO 3166-1 country code
      example: DE
...
    OfficeType:
      type: string
      pattern: '^[A-Z0-9]{1,2}$'
      description: Status of the entity in the office hierarchy
      example: HO
      # Current possible values:
      #  - HO: Head Office
      #  - MP: Main Payments Office
      #  - DB: Domestic Branch/Department
      #  - SB: Sub-Domestic Branch/Department
      #  - FB: Foreign Branch
      #  - SF: Sub-Foreign Branch
      #  - UC: Unclassified
    IbanDetailsWithIsoFlag:
      type: object
      required:
        - iban
        - country_code
        - checksum
        - bank_id
        - account_number
        - length
        - is_iso
      properties:
        iban:
          $ref: '#/components/schemas/Iban'
          # description: The IBAN that was validated
        country_code:
          $ref: '#/components/schemas/CountryCode'
          # description: The ISO 3166-1 alpha-2 code of the country of the institution/branch.
        checksum:
          $ref: '#/components/schemas/TwoDigitCode'
          # description: The checksum extracted from the IBAN
        bank_id:
          $ref: '#/components/schemas/AlphaNumStringMax30'
          # description: The BANK ID extracted from the IBAN. Its structure (bank ID only or bank ID + branch ID) is as defined by the IBAN BIC LENGTH in the IBANSTRUCTURE file.
        branch_id:
          $ref: '#/components/schemas/AlphaNumStringMax30'
          # description: Unique and unambiguous identification of a branch of a financial institution.
        account_number:
          $ref: '#/components/schemas/AlphaNumStringMax30'
          # description: The remainder of the IBAN including the account number.
        length:
          type: integer
          minimum: 5
          maximum: 34
          description: The length of the IBAN
        is_iso:
          type: boolean
          description: Whether the IBAN format for the country has been officially registered with ISO (ISO 13616)
      xml:
        name: iban
      example:
        iban: DE04500700100532013000
        country_code: DE
        checksum: '04'
        bank_id: '50070010'
        account_number: '0532013000'
        length: 22
        is_iso: true
    BicIsoDetails:
      type: object
      required:
        - bic
        - name
        - structured_address
      properties:
        bic:
          $ref: '#/components/schemas/Bic'
          # description: The BIC of the institution on which more information is requested
        office_type:
          $ref: '#/components/schemas/OfficeType'
        name:
          $ref: '#/components/schemas/InstitutionName'
          # description: Name by which an institution is known and identified
        department:
          $ref: '#/components/schemas/InstitutionDepartment'
        sub_department:
          $ref: '#/components/schemas/InstitutionDepartment'
        unstructured_address:
          $ref: '#/components/schemas/UnstructuredAddress'
        structured_address:
          $ref: '#/components/schemas/StructuredAddress'
        swift_connectivity:
          $ref: '#/components/schemas/SwiftConnectivity'
          # description: BIC connectivity
        swift_services:
          type: array
          items:
            $ref: '#/components/schemas/SwiftService'
          xml:
            wrapped: true
          description: A list of 3-character codes and names of SWIFT FIN service codes (also called value added service codes)
        swift_type:
          $ref: '#/components/schemas/SwiftSubtypeIndicator'
        utc_offset:
          $ref: '#/components/schemas/TimeOffset'
      xml:
        name: bic
      example:
        bic: 'UBSWGB2LSBI'
        office_type: 'SF'
        name: 'UBS AG LONDON BRANCH'
        department: 'UBS WARBURG INTERNATIONAL LTD, FORMERLY SBCI INVESTMENT BANKING LTD'
        sub_department: 'BQCODES: BRA'
        unstructured_address:
          address_line_1: 'SWISS BANK HOUSE'
          address_line_2: '1 HIGH TIMBER STREET'
          address_line_3: 'LONDON, EC4V 3SB, GREATER LONDON GB'
        structured_address:
          street_name: 'HIGH TIMBER STREET'
          building_number: '1'
          building_name: 'SWISS BANK HOUSE'
          post_code: 'EC4V 3SB'
          town_name: 'LONDON'
          country_subdivision_name: 'GREATER LONDON'
          country_code: 'GB'
          country_name: 'UNITED KINGDOM'
        swift_connectivity:
          fin: true
          finplus:
            future: 'ou=sbi,o=ubswgb2l,o=swift'
            live: 'ou=sbi,o=ubswgb2l,o=swift'
            pilot: 'ou=sbi,o=ubswgb2l,o=swift'
          interact: false
          fileact: false
        swift_services:
          - code: 'FIN'
            name: 'Many-to-many FIN payment service'
        utc_offset: "±00:00"
...